修复类型转换问题以确保计算正确性
修复了 `Matrix4x4.CreatePerspectiveFieldOfView` 方法中 `window.FramebufferSize.X / (float)window.FramebufferSize.Y` 的类型转换问题，改为 `window.FramebufferSize.X / window.FramebufferSize.Y`，以确保计算的正确性。

+++ Source/Tests/Tests.GLTFScene/Program.cs
@@ -275,7 +275,7 @@ internal sealed unsafe class Program
 
         UBO ubo = new()
         {
-            Projection = Matrix4x4.CreatePerspectiveFieldOfView(MathF.PI / 4, window.FramebufferSize.X / (float)window.FramebufferSize.Y, 0.1f, 1000.0f),
+            Projection = Matrix4x4.CreatePerspectiveFieldOfView(MathF.PI / 4, window.FramebufferSize.X / window.FramebufferSize.Y, 0.1f, 1000.0f),
             View = Matrix4x4.CreateLookAt(new Vector3(7.8f, 2.1f, 0.0f), Vector3.Zero, Vector3.UnitY),
             Model = _nodes[0].LocalTransform,
             LightPos = Vector4.Transform(new Vector4(0.0f, 2.5f, 0.0f, 1.0f), Matrix4x4.CreateRotationX(MathF.Sin(e.TotalTime))),
